requirements.txt
streamlit
moviepy==1.0.3
openai-whisper @ git+https://github.com/openai/whisper.git@25639fc17ddc013d56c594bfbf7644f2185fad84
torch==2.4.1 
torchaudio==2.4.1
torchvision==0.19.1
llama-index-core==0.11.15
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-groq==0.2.0
llama-index-experimental==0.4.0
llama-index-readers-file==0.2.2
llama-index-postprocessor-cohere-rerank==0.2.1
pydub==0.25.1
ffmpeg-python
librosa
os
numpy